+++ src/chains/Base/Common.sol
@@ -8,7 +8,6 @@ import {DodoV2, IDodoV2} from "../../core/DodoV2.sol";
 import {MaverickV2, IMaverickV2Pool} from "../../core/MaverickV2.sol";
 import {UniswapV4} from "../../core/UniswapV4.sol";
 import {IPoolManager} from "../../core/UniswapV4Types.sol";
-import {EulerSwap, IEVC, IEulerSwap} from "../../core/EulerSwap.sol";
 import {BalancerV3} from "../../core/BalancerV3.sol";
 import {PancakeInfinity} from "../../core/PancakeInfinity.sol";
 import {
@@ -25,7 +24,6 @@ import {FreeMemory} from "../../utils/FreeMemory.sol";
 import {FastLogic} from "../../utils/FastLogic.sol";
 
 import {ISettlerActions} from "../../ISettlerActions.sol";
-import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 import {revertUnknownForkId} from "../../core/SettlerErrors.sol";
 
 import {
@@ -72,7 +70,6 @@ abstract contract BaseMixin is
     UniswapV4,
     BalancerV3,
     PancakeInfinity,
-    //EulerSwap,
     Renegade,
     Bebop,
     Hanji
@@ -117,13 +114,6 @@ abstract contract BaseMixin is
             } else { // if (action == uint32(ISettlerActions.PANCAKE_INFINITY.selector))
                 sellToPancakeInfinity(recipient, sellToken, ppm, feeOnTransfer, hashMul, hashMod, fills, amountOutMin);
             }
-        /*
-        } else if (action == uint32(ISettlerActions.EULERSWAP.selector)) {
-            (address recipient, IERC20 sellToken, uint256 ppm, IEulerSwap pool, bool zeroForOne, uint256 amountOutMin) =
-                abi.decode(data, (address, IERC20, uint256, IEulerSwap, bool, uint256));
-
-            sellToEulerSwap(recipient, sellToken, ppm, pool, zeroForOne, amountOutMin);
-        */
         } else if (action == uint32(ISettlerActions.MAVERICKV2.selector)) {
             (
                 address recipient,
@@ -264,12 +254,6 @@ abstract contract BaseMixin is
     function _PANCAKE_INFINITY_BIN_MANAGER() internal pure override returns (address) {
         return pancakeInfinityBinManager;
     }
-
-    /*
-    function _EVC() internal pure override returns (IEVC) {
-        return IEVC(0x5301c7dD20bD945D2013b48ed0DEE3A284ca8989);
-    }
-    */
 
     function _fallback(bytes calldata data)
         internal