+++ src/chains/Berachain/Intent.sol
@@ -4,15 +4,12 @@ pragma solidity =0.8.34;
 import {BerachainSettlerMetaTxn} from "./MetaTxn.sol";
 import {SettlerIntent} from "../../SettlerIntent.sol";
 
-import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
-import {ISettlerActions} from "../../ISettlerActions.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../../SettlerAbstract.sol";
 import {SettlerBase} from "../../SettlerBase.sol";
 import {SettlerMetaTxn} from "../../SettlerMetaTxn.sol";
-import {SettlerIntent} from "../../SettlerIntent.sol";
 import {AbstractContext, Context} from "../../Context.sol";
 import {Permit2PaymentAbstract} from "../../core/Permit2PaymentAbstract.sol";
 import {Permit2PaymentMetaTxn} from "../../core/Permit2Payment.sol";
@@ -30,14 +27,6 @@ contract BerachainSettlerIntent is SettlerIntent, BerachainSettlerMetaTxn {
         bytes calldata sig
     ) public override(SettlerIntent, SettlerMetaTxn) returns (bool) {
         return super.executeMetaTxn(slippage, actions, bytes32(0), msgSender, sig);
-    }
-
-    function _dispatch(uint256 i, uint256 action, bytes calldata data, AllowedSlippage memory slippage)
-        internal
-        override(BerachainSettlerMetaTxn, SettlerBase)
-        returns (bool)
-    {
-        return super._dispatch(i, action, data, slippage);
     }
 
     function _isForwarded() internal view override(AbstractContext, Context, SettlerIntent) returns (bool) {